requirements.txt
# ============================================
# EchoMind - Requirements
# ============================================

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1

# Machine Learning
prophet==1.1.5
scikit-learn==1.3.2
joblib==1.3.2

# Data Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# API & Validation
marshmallow==3.20.1
python-dotenv==1.0.0
email-validator==2.1.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Production Server
gunicorn==21.2.0

# Visualization (for training scripts)
matplotlib==3.8.2
seaborn==0.13.0